+++ src/utils/common/add-class.js
@@ -1,0 +1,8 @@
+export const addClass = (el, className) => {
+  const element = el || {};
+  if (element.classList) {
+    element.classList.add(className);
+  } else {
+    element.className += ` ${className}`;
+  }
+};