+++ src/utils/container-elements/create-magnifier-icon.js
@@ -1,0 +1,7 @@
+export const createMagnifierIcon = () => {
+  const magnifyIcon = document.createElement('div');
+
+  magnifyIcon.className = 'cloudimage-360-magnifier-icon';
+
+  return magnifyIcon;
+};