+++ src/utils/hotspot/elements/create-modal-description.js
@@ -1,0 +1,8 @@
+export const createModalDescription = (description) => {
+  const modalDescription = document.createElement('p');
+
+  modalDescription.innerText = description;
+  modalDescription.className = 'cloudimage-360-modal-description';
+
+  return modalDescription;
+};