+++ src/utils/hotspot/get-hotspot-orientation.js
@@ -1,0 +1,12 @@
+export const getHotspotOriantaion = (orientation) => {
+  switch (orientation.toLowerCase()) {
+    case 'x-axis':
+      return 'x';
+
+    case 'y-axis':
+      return 'y';
+
+    default:
+      return 'x';
+  }
+};