+++ src/utils/hotspot/get-hotspot-popup-node.js
@@ -1,0 +1,10 @@
+export const getHotspotPopupNode = (anchorId, open) => {
+  const hotspotPopupSelector = `[data-hotspot-popup-id=${anchorId}]`;
+  const hotspotPopup = document.querySelector(hotspotPopupSelector);
+
+  if (open) {
+    hotspotPopup.setAttribute('data-show', '');
+  }
+
+  return hotspotPopup;
+};