+++ src/utils/load-images/load-image.js
@@ -1,0 +1,13 @@
+export const loadImage = (src, index, srcConfig) => {
+  const { lazyload, lazySelector, fullscreenView } = srcConfig || {};
+  const image = new Image();
+
+  if (lazyload && !fullscreenView) {
+    image.setAttribute('data-src', src);
+    image.className = image.className.length ? `${image.className} ${lazySelector}` : lazySelector;
+  } else {
+    image.src = src;
+  }
+
+  return image;
+};