+++ src/utils/responsive/get-container-responsive-height.js
@@ -1,0 +1,11 @@
+export const getContainerResponsiveHeight = (container, width, containerHeight) => {
+  if (containerHeight) {
+    if (width < containerHeight) {
+      return width;
+    }
+
+    return containerHeight;
+  }
+
+  return container.offsetWidth;
+};